+++ example/lib/main.dart
@@ -24,18 +24,25 @@ class _MazeScreenState extends State<MazeScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        body: SafeArea(
-            child: Maze(
+      body: SafeArea(
+        child: Center(
+          child: TextButton(
+            child: Text('Show Maze'),
+            onPressed: () => MazeMinigame.show(
+              MazeProperties(
+                context: context,
                 player: MazeItem(
                     'https://image.flaticon.com/icons/png/512/808/808433.png',
                     ImageType.network),
-                columns: 6,
-                rows: 12,
-                wallThickness: 4.0,
-                wallColor: Theme.of(context).primaryColor,
-                finish: MazeItem(
+                finishImage: MazeItem(
                     'https://image.flaticon.com/icons/png/512/1506/1506339.png',
                     ImageType.network),
-                onFinish: () => print('Hi from finish line!'))));
+                onFinish: () => print('Hi from finish line!'),
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
   }
 }