+++ lib/maze.dart
@@ -1,2 +1,3 @@
 export './src/maze_widget.dart';
 export './src/models/item.dart';
+export './src/models/maze_properties.dart';