+++ lib/src/maze_widget.dart
@@ -1,130 +1,142 @@
+// ignore_for_file: public_member_api_docs
+
 import 'dart:async';
 import 'dart:ui' as ui;
+import 'package:back_button_interceptor/back_button_interceptor.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:http/http.dart' as http;
+import 'package:maze/src/models/maze_properties.dart';
 import 'package:universal_io/io.dart';
 import 'maze_painter.dart';
 import 'models/item.dart';
 
+class MazeMinigame {
+  static void show(MazeProperties mazeProperties) {
+    Future.delayed(
+      const Duration(milliseconds: 300),
+      () => showGeneralDialog(
+        barrierDismissible: false,
+        barrierColor: Colors.transparent,
+        context: mazeProperties.context,
+        transitionDuration: const Duration(milliseconds: 500),
+        transitionBuilder: (context, anim1, anim2, child) {
+          return SlideTransition(
+            position: Tween(begin: const Offset(0, 1), end: const Offset(0, 0))
+                .animate(anim1),
+            child: child,
+          );
+        },
+        pageBuilder: (context, anim1, anim2) {
+          return Align(
+            alignment: Alignment.center,
+            child: Dialog(
+              backgroundColor: Colors.transparent,
+              child: _Maze(mazeProperties: mazeProperties),
+            ),
+          );
+        },
+      ),
+    );
+  }
+}
+
 ///Maze
 ///
 ///Create a simple but powerfull maze game
-///You can customize [wallColor], [wallThickness],
-///[columns] and [rows]. A [player] is required and also
-///you can pass a List of [checkpoints] and you will be notified
-///if the player pass through a checkout at [onCheckpoint]
-class Maze extends StatefulWidget {
+///You can customize [mazeProperties]
+class _Maze extends StatefulWidget {
   ///Default constructor
-  Maze({
-    required this.player,
-    this.checkpoints = const [],
-    this.columns = 10,
-    this.finish,
-    this.height,
-    this.loadingWidget,
-    this.onCheckpoint,
-    this.onFinish,
-    this.rows = 7,
-    this.wallColor = Colors.black,
-    this.wallThickness = 3.0,
-    this.width,
-  });
-
-  ///List of checkpoints
-  final List<MazeItem> checkpoints;
-
-  ///Columns of the maze
-  final int columns;
-
-  ///The finish image
-  final MazeItem? finish;
-
-  ///Height of the maze
-  final double? height;
-
-  ///A widget to show while loading all
-  final Widget? loadingWidget;
-
-  ///Callback when the player pass through a checkpoint
-  final Function(int)? onCheckpoint;
-
-  ///Callback when the player reach finish
-  final Function()? onFinish;
-
-  ///The main player
-  final MazeItem player;
-
-  ///Rows of the maze
-  final int rows;
-
-  ///Wall color
-  final Color? wallColor;
-
-  ///Wall thickness
-  ///
-  ///Default: 3.0
-  final double? wallThickness;
-
-  ///Width of the maze
-  final double? width;
+  _Maze({required this.mazeProperties});
+  // Maze properties
+  final MazeProperties mazeProperties;
 
   @override
   _MazeState createState() => _MazeState();
 }
 
-class _MazeState extends State<Maze> {
+class _MazeState extends State<_Maze> {
   bool _loaded = false;
   late MazePainter _mazePainter;
+  bool myInterceptor(bool stopDefaultButtonEvent, RouteInfo info) {
+    return true;
+  }
+
+  final FocusNode _focusNode = FocusNode();
+  int lastKbEvent = DateTime.now().millisecondsSinceEpoch;
 
   @override
   void initState() {
     super.initState();
+    BackButtonInterceptor.add(myInterceptor);
     setUp();
   }
 
+  @override
+  void dispose() {
+    BackButtonInterceptor.remove(myInterceptor);
+    super.dispose();
+  }
+
   void setUp() async {
-    final playerImage = await _itemToImage(widget.player);
-    final checkpoints = await Future.wait(
-        widget.checkpoints.map((c) async => await _itemToImage(c)));
-    final finishImage =
-        widget.finish != null ? await _itemToImage(widget.finish!) : null;
+    final playerImage = await _itemToImage(widget.mazeProperties.player);
+    final checkpoints = await Future.wait(widget.mazeProperties.checkpoints
+        .map((c) async => await _itemToImage(c)));
+    final finishImage = widget.mazeProperties.finishImage != null
+        ? await _itemToImage(widget.mazeProperties.finishImage!)
+        : null;
 
     _mazePainter = MazePainter(
+      context: context,
       checkpointsImages: checkpoints,
-      columns: widget.columns,
+      columns: widget.mazeProperties.columns,
       finishImage: finishImage,
-      onCheckpoint: widget.onCheckpoint,
-      onFinish: widget.onFinish,
+      onCheckpoint: widget.mazeProperties.onCheckpoint,
+      onFinish: widget.mazeProperties.onFinish,
       playerImage: playerImage,
-      rows: widget.rows,
-      wallColor: widget.wallColor ?? Colors.black,
-      wallThickness: widget.wallThickness ?? 4.0,
+      rows: widget.mazeProperties.rows,
+      wallColor: widget.mazeProperties.wallColor ?? Colors.black,
+      wallThickness: widget.mazeProperties.wallThickness ?? 4.0,
     );
     setState(() => _loaded = true);
   }
 
   @override
   Widget build(BuildContext context) {
-    return Container(child: Builder(builder: (context) {
-      if (_loaded) {
-        return GestureDetector(
-            onVerticalDragUpdate: (info) =>
-                _mazePainter.updatePosition(info.localPosition),
-            child: CustomPaint(
-                painter: _mazePainter,
-                size: Size(widget.width ?? context.width,
-                    widget.height ?? context.height)));
-      } else {
-        if (widget.loadingWidget != null) {
-          return widget.loadingWidget!;
-        } else {
-          return const Center(
-            child: CircularProgressIndicator(),
-          );
-        }
-      }
-    }));
+    return Material(
+      color: widget.mazeProperties.backgroundColor ??
+          Colors.white.withOpacity(0.8),
+      child: Container(
+        child: Builder(
+          builder: (context) {
+            if (_loaded) {
+              return RawKeyboardListener(
+                autofocus: true,
+                focusNode: _focusNode,
+                onKey: _handleKeyEvent,
+                child: GestureDetector(
+                  onVerticalDragUpdate: (info) =>
+                      _mazePainter.updatePosition(info.localPosition, null),
+                  child: CustomPaint(
+                    painter: _mazePainter,
+                    size: Size(widget.mazeProperties.width ?? context.width,
+                        widget.mazeProperties.height ?? context.height),
+                  ),
+                ),
+              );
+            } else {
+              if (widget.mazeProperties.loadingWidget != null) {
+                return widget.mazeProperties.loadingWidget!;
+              } else {
+                return const Center(
+                  child: CircularProgressIndicator(),
+                );
+              }
+            }
+          },
+        ),
+      ),
+    );
   }
 
   Future<ui.Image> _itemToImage(MazeItem item) {
@@ -161,6 +173,29 @@ class _MazeState extends State<Maze> {
     ui.decodeImageFromList(
         response.bodyBytes.buffer.asUint8List(), completer.complete);
     return completer.future;
+  }
+
+  void _handleKeyEvent(RawKeyEvent event) {
+    final currentKbEvent = DateTime.now().millisecondsSinceEpoch;
+    final allowKbEvent = (currentKbEvent - lastKbEvent) > 150;
+
+    if (allowKbEvent) {
+      if (event.logicalKey == LogicalKeyboardKey.keyA ||
+          event.logicalKey == LogicalKeyboardKey.arrowLeft) {
+        _mazePainter.updatePosition(Offset.zero, Direction.left);
+      } else if (event.logicalKey == LogicalKeyboardKey.keyW ||
+          event.logicalKey == LogicalKeyboardKey.arrowUp) {
+        _mazePainter.updatePosition(Offset.zero, Direction.up);
+      } else if (event.logicalKey == LogicalKeyboardKey.keyD ||
+          event.logicalKey == LogicalKeyboardKey.arrowRight) {
+        _mazePainter.updatePosition(Offset.zero, Direction.right);
+      } else if (event.logicalKey == LogicalKeyboardKey.keyS ||
+          event.logicalKey == LogicalKeyboardKey.arrowDown) {
+        _mazePainter.updatePosition(Offset.zero, Direction.down);
+      }
+
+      lastKbEvent = currentKbEvent;
+    }
   }
 }